+++ packages/lib/src/model/modelSymbols.ts
@@ -1,4 +1,3 @@
-export const modelPropertiesSymbol = Symbol("modelProperties")
 export const modelDataTypeCheckerSymbol = Symbol("modelDataTypeChecker")
 export const modelInitializersSymbol = Symbol("modelInitializers")
 export const modelUnwrappedClassSymbol = Symbol("modelUnwrappedClass")