+++ packages/lib/src/model/prop.ts
@@ -1,7 +1,18 @@
 import { O } from "ts-toolbelt"
 import { PropTransform } from "../propTransform/propTransform"
 import { LateTypeChecker, TypeChecker } from "../typeChecking/TypeChecker"
+import { isObject } from "../utils"
 import { IsOptionalValue } from "../utils/types"
+
+/**
+ * Model property options.
+ */
+export interface ModelPropOptions {
+  /**
+   * Set to `true` to automatically wrap the property setter in a model action (defaults to `false`).
+   */
+  readonly setterAction?: boolean
+}
 
 /**
  * @ignore
@@ -23,10 +34,12 @@ export interface ModelProp<
   $instanceValueType: TInstanceValue
   $instanceCreationValueType: TInstanceCreationValue
   $isOptional: TIsOptional
+
   defaultFn: (() => TPropValue) | typeof noDefaultValue
   defaultValue: TPropValue | typeof noDefaultValue
   typeChecker: TypeChecker | LateTypeChecker | undefined
   transform: PropTransform<any, any> | undefined
+  options: ModelPropOptions
 }
 
 /**
@@ -90,20 +103,6 @@ export type OptionalModelProp<TPropValue, TInstanceValue = TPropValue> = ModelPr
 >
 
 /**
- * Defines a model property with no default value.
- *
- * Example:
- * ```ts
- * x: prop<number>() // a required number
- * x: prop<number | undefined>() // an optional number, which defaults to undefined
- * ```
- *
- * @typeparam TValue Value type.
- * @returns
- */
-export function prop<TValue>(): MaybeOptionalModelProp<TValue>
-
-/**
  * Defines a model property, with an optional function to generate a default value
  * if the input snapshot / model creation data is `null` or `undefined`.
  *
@@ -115,9 +114,13 @@ export function prop<TValue>(): MaybeOptionalModelProp<TValue>
  *
  * @typeparam TValue Value type.
  * @param defaultFn Default value generator function.
+ * @param options Model property options.
  * @returns
  */
-export function prop<TValue>(defaultFn: () => TValue): OptionalModelProp<TValue>
+export function prop<TValue>(
+  defaultFn: () => TValue,
+  options?: ModelPropOptions
+): OptionalModelProp<TValue>
 
 /**
  * Defines a model property, with an optional default value
@@ -132,12 +135,51 @@ export function prop<TValue>(defaultFn: () => TValue): OptionalModelProp<TValue>
  *
  * @typeparam TValue Value type.
  * @param defaultValue Default primitive value.
+ * @param options Model property options.
  * @returns
  */
-export function prop<TValue>(defaultValue: OnlyPrimitives<TValue>): OptionalModelProp<TValue>
+export function prop<TValue>(
+  defaultValue: OnlyPrimitives<TValue>,
+  options?: ModelPropOptions
+): OptionalModelProp<TValue>
 
-export function prop<TValue>(def?: any): ModelProp<TValue, any, any> {
-  const hasDefaultValue = arguments.length > 0
+/**
+ * Defines a model property with no default value.
+ *
+ * Example:
+ * ```ts
+ * x: prop<number>() // a required number
+ * x: prop<number | undefined>() // an optional number, which defaults to undefined
+ * ```
+ *
+ * @typeparam TValue Value type.
+ * @param options Model property options.
+ * @returns
+ */
+export function prop<TValue>(options?: ModelPropOptions): MaybeOptionalModelProp<TValue>
+
+export function prop<TValue>(arg1?: any, arg2?: any): ModelProp<TValue, any, any> {
+  let def: any
+  let opts: ModelPropOptions = {}
+  let hasDefaultValue = false
+
+  if (arguments.length >= 2) {
+    // default, options
+    def = arg1
+    hasDefaultValue = true
+    opts = { ...arg2 }
+  } else if (arguments.length === 1) {
+    // default | options
+    if (isObject(arg1)) {
+      // options
+      opts = { ...arg1 }
+    } else {
+      // default
+      def = arg1
+      hasDefaultValue = true
+    }
+  }
+
   const isDefFn = typeof def === "function"
 
   return {
@@ -151,5 +193,6 @@ export function prop<TValue>(def?: any): ModelProp<TValue, any, any> {
     defaultValue: hasDefaultValue && !isDefFn ? def : noDefaultValue,
     typeChecker: undefined,
     transform: undefined,
+    options: opts,
   }
 }