+++ packages/lib/src/parent/coreObjectChildren.ts
@@ -1,5 +1,6 @@
 import { action, createAtom, IAtom, observable, ObservableSet } from "mobx"
 import { AnyModel } from "../model/BaseModel"
+import { modelIdKey, modelTypeKey } from "../model/metadata"
 import { isModel } from "../model/utils"
 import { fastGetParent } from "./path"
 
@@ -58,7 +59,7 @@ function addNodeToDeepLists(
 ) {
   deep.add(node)
   if (isModel(node)) {
-    deepByModelTypeAndId.set(byModelTypeAndIdKey(node.$modelType, node.$modelId), node)
+    deepByModelTypeAndId.set(byModelTypeAndIdKey(node[modelTypeKey], node[modelIdKey]), node)
   }
 }