+++ packages/lib/src/utils/ModelPool.ts
@@ -1,4 +1,5 @@
 import { AnyModel } from "../model/BaseModel"
+import { modelIdKey, modelTypeKey } from "../model/metadata"
 import { isModelSnapshot } from "../model/utils"
 import { dataObjectParent } from "../parent/core"
 import { byModelTypeAndIdKey, getDeepObjectChildren } from "../parent/coreObjectChildren"
@@ -22,6 +23,6 @@ export class ModelPool {
       return undefined
     }
 
-    return this.findModelByTypeAndId(sn.$modelType, sn.$modelId)
+    return this.findModelByTypeAndId(sn[modelTypeKey], sn[modelIdKey])
   }
 }